Rename StyledDiv to stylesFormat
