+++ client/components/Story.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { StyledDesc, NavHeader, SectionHeader, Caption, Img, FontFaces } from '../StyledDiv';
+import { StyledDesc, NavHeader, SectionHeader, Caption, Img, FontFaces } from '../stylesFormat';
 
 const Story = ({story}) => {
   const { textTop, imageMiddle, imageMiddleCaption, textBottom } = story;
